fix(replay): depend on Salesforce Extension Pack
Declare salesforce.salesforcedx-vscode as an extensionDependency and remove manual Replay warm-up/install prompts.

Also increases webview Jest timeout to avoid flakes in slower CI.
