+++ scripts/gen-nls.cjs
@@ -23,10 +23,8 @@ const en = {
   openError: 'Failed to open log: ',
   replayError: 'Failed to launch Apex Replay Debugger: ',
   replayStarting: 'Starting Apex Replay Debugger…',
-  replayMissingExtMessage: 'Salesforce Extension Pack (includes Apex Replay Debugger) is required to replay logs.',
-  replayMissingExtInstall: 'Install',
-  replayMissingExtOpen: 'Open Extensions',
-  replayPackAdvice: 'Tip: To use Apex Replay, install the Salesforce Extension Pack (salesforce.salesforcedx-vscode).',
+  replayMissingExtMessage:
+    'Apex Replay Debugger is unavailable. Ensure the Salesforce Extension Pack (salesforce.salesforcedx-vscode) is installed and enabled.',
   tailSelectDebugLevel: 'Select a debug level',
   tailHardStop: 'Tail stopped after 30 minutes.',
   tailSavedTo: 'Saved to {0}',
@@ -49,10 +47,8 @@ const ptBr = {
   openError: 'Falha ao abrir o log: ',
   replayError: 'Falha ao iniciar o Apex Replay Debugger: ',
   replayStarting: 'Iniciando o Apex Replay Debugger…',
-  replayMissingExtMessage: 'Para reproduzir logs é necessário o Salesforce Extension Pack (inclui o Apex Replay Debugger).',
-  replayMissingExtInstall: 'Instalar',
-  replayMissingExtOpen: 'Abrir Extensões',
-  replayPackAdvice: 'Dica: Para usar o Apex Replay, instale o Salesforce Extension Pack (salesforce.salesforcedx-vscode).',
+  replayMissingExtMessage:
+    'Apex Replay Debugger não está disponível. Garanta que o Salesforce Extension Pack (salesforce.salesforcedx-vscode) esteja instalado e habilitado.',
   tailSelectDebugLevel: 'Selecione um nível de depuração',
   tailHardStop: 'Tail parado após 30 minutos.',
   tailSavedTo: 'Salvo em {0}',
